requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.3; python_version >= '3.8'
aiohttp==3.11.6; python_version >= '3.9'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.6.2.post1; python_version >= '3.9'
attrs==24.2.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
colorclass==2.2.2; python_version >= '2.6'
compressed-rtf==1.0.6
cryptography==43.0.3; python_version >= '3.7'
easygui==0.98.3
ebcdic==1.1.1
eml-parser==2.0.0; python_version >= '3.9'
extract-msg==0.52.0; python_version >= '3.8'
frozenlist==1.5.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
langchain==0.3.7; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.19; python_version >= '3.9' and python_version < '4.0'
langchain-ollama==0.2.0; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.3.2; python_version >= '3.9' and python_version < '4.0'
langsmith==0.1.144; python_full_version >= '3.8.1' and python_version < '4.0'
lark==1.1.9; python_version >= '3.6'
msoffcrypto-tool==5.4.2; python_version >= '3.8' and python_version < '4.0'
multidict==6.1.0; python_version >= '3.8'
numpy==1.26.4; python_version >= '3.9'
olefile==0.47; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
oletools==0.60.2
ollama==0.3.3; python_version >= '3.8' and python_version < '4.0'
orjson==3.10.11; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pcodedmp==1.2.6
propcache==0.2.0; python_version >= '3.8'
publicsuffixlist==1.0.2.20241121; python_version >= '3.5'
pycparser==2.22; python_version >= '3.8'
pydantic==2.10.0; python_version >= '3.8'
pydantic-core==2.27.0; python_version >= '3.8'
pyparsing==3.2.0; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-magic==0.4.27; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyyaml==6.0.2; python_version >= '3.8'
red-black-tree-mod==1.20
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rtfde==0.1.2; python_version ~= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
sqlalchemy==2.0.36; python_version >= '3.7'
tenacity==9.0.0; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
tzlocal==5.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
yarl==1.17.2; python_version >= '3.9'
